# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clothes E-Commerce API",
        "description": "Comprehensive API Documentation for the full e-commerce system (Customer & Admin routes).",
        "contact": {
            "name": "Figo Team",
            "email": "admin@clothes.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/admin/api/categories": {
            "get": {
                "tags": [
                    "Admin - Category"
                ],
                "summary": "Get list of Categorys",
                "operationId": "4f443bd9d31392f52ded9aa72ddf7770",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin - Category"
                ],
                "summary": "Create/Process Category (store)",
                "operationId": "222ed03b8db4f944c13a3a66cd0bec21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "icon"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/admin/api/categories/{id}": {
            "put": {
                "tags": [
                    "Admin - Category"
                ],
                "summary": "Update Category (update)",
                "operationId": "d53768c239c13723bd2dd8434ba49adb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "icon"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin - Category"
                ],
                "summary": "Delete Category",
                "operationId": "4a34be497794781f260874a99fa51938",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/admin/api/inventory/{id}/adjust": {
            "put": {
                "tags": [
                    "Admin - Inventory"
                ],
                "summary": "Update Inventory (adjust)",
                "operationId": "781f7eae022e7ba35c513f1475a88bd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "reason"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/admin/api/orders": {
            "get": {
                "tags": [
                    "Admin - Order"
                ],
                "summary": "Get list of Orders",
                "operationId": "3611b1660b68f82da6b186b990bae4d6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/admin/api/orders/{id}": {
            "get": {
                "tags": [
                    "Admin - Order"
                ],
                "summary": "Get specific Order",
                "operationId": "7f4c1b88917abe76e0fbd238f35ffcc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/admin/api/orders/{id}/status": {
            "put": {
                "tags": [
                    "Admin - Order"
                ],
                "summary": "Update Order (updateStatus)",
                "operationId": "0c68029c64b5c9c32264d1398d50836e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format **Bearer <token>**",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}